Create model importer and move input file mappings to file
