+++ src/meshpy/four_c/input_file.py
@@ -33,63 +33,28 @@ from typing import Dict as _Dict
 from typing import List as _List
 from typing import Optional as _Optional
 from typing import Tuple as _Tuple
-from typing import TypeVar as _TypeVar
-from typing import Union as _Union
 
 import yaml as _yaml
 
-import meshpy.core.conf as _conf
 from meshpy.core.boundary_condition import BoundaryCondition as _BoundaryCondition
-from meshpy.core.boundary_condition import (
-    BoundaryConditionBase as _BoundaryConditionBase,
-)
 from meshpy.core.conf import mpy as _mpy
 from meshpy.core.container import ContainerBase as _ContainerBase
-from meshpy.core.coupling import Coupling as _Coupling
-from meshpy.core.element import Element as _Element
 from meshpy.core.geometry_set import GeometryName as _GeometryName
-from meshpy.core.geometry_set import GeometrySetNodes as _GeometrySetNodes
 from meshpy.core.mesh import Mesh as _Mesh
-from meshpy.core.node import Node as _Node
 from meshpy.core.nurbs_patch import NURBSPatch as _NURBSPatch
+from meshpy.four_c.input_file_mappings import (
+    boundary_condition_names as _boundary_condition_names,
+)
+from meshpy.four_c.input_file_mappings import geometry_set_names as _geometry_set_names
 from meshpy.four_c.yaml_dumper import MeshPyDumper as _MeshPyDumper
 from meshpy.utils.environment import cubitpy_is_available as _cubitpy_is_available
 from meshpy.utils.environment import fourcipp_is_available as _fourcipp_is_available
-
-# necessary to allow for type hint of from_4C_yaml for python version <3.11
-# can be replaced with _Self in python 3.11
-T = _TypeVar("T", bound="InputFile")
 
 if _cubitpy_is_available():
     import cubitpy as _cubitpy
 
 
-def _boundary_condition_from_dict(
-    geometry_set: _GeometrySetNodes,
-    bc_key: _Union[_conf.BoundaryCondition, str],
-    data: _Dict,
-) -> _BoundaryConditionBase:
-    """This function acts as a factory and creates the correct boundary
-    condition object from a dictionary parsed from an input file."""
-
-    del data["E"]
-
-    if bc_key in (
-        _mpy.bc.dirichlet,
-        _mpy.bc.neumann,
-        _mpy.bc.locsys,
-        _mpy.bc.beam_to_solid_surface_meshtying,
-        _mpy.bc.beam_to_solid_surface_contact,
-        _mpy.bc.beam_to_solid_volume_meshtying,
-    ) or isinstance(bc_key, str):
-        return _BoundaryCondition(geometry_set, data, bc_type=bc_key)
-    elif bc_key is _mpy.bc.point_coupling:
-        return _Coupling(geometry_set, bc_key, data, check_overlapping_nodes=False)
-    else:
-        raise ValueError("Got unexpected boundary condition!")
-
-
-def _get_geometry_set_indices_from_section(
+def get_geometry_set_indices_from_section(
     section_list: _List, *, append_node_ids: bool = True
 ) -> _Dict:
     """Return a dictionary with the geometry set ID as keys and the node IDs as
@@ -116,96 +81,11 @@ def _get_geometry_set_indices_from_section(
     return geometry_set_dict
 
 
-def _get_yaml_geometry_sets(
-    nodes: _List[_Node], geometry_key: _conf.Geometry, section_list: _List
-) -> _Dict[int, _GeometrySetNodes]:
-    """Add sets of points, lines, surfaces or volumes to the object."""
-
-    # Create the individual geometry sets. The nodes are still integers at this
-    # point. They have to be converted to links to the actual nodes later on.
-    geometry_set_dict = _get_geometry_set_indices_from_section(section_list)
-    geometry_sets_in_this_section = {}
-    for geometry_set_id, node_ids in geometry_set_dict.items():
-        geometry_sets_in_this_section[geometry_set_id] = _GeometrySetNodes(
-            geometry_key, nodes=[nodes[node_id] for node_id in node_ids]
-        )
-    return geometry_sets_in_this_section
-
-
 class InputFile:
     """An item that represents a complete 4C input file."""
 
-    # Define the names of sections and boundary conditions in the input file.
-    geometry_set_names = {
-        _mpy.geo.point: "DNODE-NODE TOPOLOGY",
-        _mpy.geo.line: "DLINE-NODE TOPOLOGY",
-        _mpy.geo.surface: "DSURF-NODE TOPOLOGY",
-        _mpy.geo.volume: "DVOL-NODE TOPOLOGY",
-    }
-    boundary_condition_names = {
-        (_mpy.bc.dirichlet, _mpy.geo.point): "DESIGN POINT DIRICH CONDITIONS",
-        (_mpy.bc.dirichlet, _mpy.geo.line): "DESIGN LINE DIRICH CONDITIONS",
-        (_mpy.bc.dirichlet, _mpy.geo.surface): "DESIGN SURF DIRICH CONDITIONS",
-        (_mpy.bc.dirichlet, _mpy.geo.volume): "DESIGN VOL DIRICH CONDITIONS",
-        (_mpy.bc.locsys, _mpy.geo.point): "DESIGN POINT LOCSYS CONDITIONS",
-        (_mpy.bc.locsys, _mpy.geo.line): "DESIGN LINE LOCSYS CONDITIONS",
-        (_mpy.bc.locsys, _mpy.geo.surface): "DESIGN SURF LOCSYS CONDITIONS",
-        (_mpy.bc.locsys, _mpy.geo.volume): "DESIGN VOL LOCSYS CONDITIONS",
-        (_mpy.bc.neumann, _mpy.geo.point): "DESIGN POINT NEUMANN CONDITIONS",
-        (_mpy.bc.neumann, _mpy.geo.line): "DESIGN LINE NEUMANN CONDITIONS",
-        (_mpy.bc.neumann, _mpy.geo.surface): "DESIGN SURF NEUMANN CONDITIONS",
-        (_mpy.bc.neumann, _mpy.geo.volume): "DESIGN VOL NEUMANN CONDITIONS",
-        (
-            _mpy.bc.moment_euler_bernoulli,
-            _mpy.geo.point,
-        ): "DESIGN POINT MOMENT EB CONDITIONS",
-        (
-            _mpy.bc.beam_to_solid_volume_meshtying,
-            _mpy.geo.line,
-        ): "BEAM INTERACTION/BEAM TO SOLID VOLUME MESHTYING LINE",
-        (
-            _mpy.bc.beam_to_solid_volume_meshtying,
-            _mpy.geo.volume,
-        ): "BEAM INTERACTION/BEAM TO SOLID VOLUME MESHTYING VOLUME",
-        (
-            _mpy.bc.beam_to_solid_surface_meshtying,
-            _mpy.geo.line,
-        ): "BEAM INTERACTION/BEAM TO SOLID SURFACE MESHTYING LINE",
-        (
-            _mpy.bc.beam_to_solid_surface_meshtying,
-            _mpy.geo.surface,
-        ): "BEAM INTERACTION/BEAM TO SOLID SURFACE MESHTYING SURFACE",
-        (
-            _mpy.bc.beam_to_solid_surface_contact,
-            _mpy.geo.line,
-        ): "BEAM INTERACTION/BEAM TO SOLID SURFACE CONTACT LINE",
-        (
-            _mpy.bc.beam_to_solid_surface_contact,
-            _mpy.geo.surface,
-        ): "BEAM INTERACTION/BEAM TO SOLID SURFACE CONTACT SURFACE",
-        (_mpy.bc.point_coupling, _mpy.geo.point): "DESIGN POINT COUPLING CONDITIONS",
-        (
-            _mpy.bc.beam_to_beam_contact,
-            _mpy.geo.line,
-        ): "BEAM INTERACTION/BEAM TO BEAM CONTACT CONDITIONS",
-        (
-            _mpy.bc.point_coupling_penalty,
-            _mpy.geo.point,
-        ): "DESIGN POINT PENALTY COUPLING CONDITIONS",
-        (
-            "DESIGN SURF MORTAR CONTACT CONDITIONS 3D",
-            _mpy.geo.surface,
-        ): "DESIGN SURF MORTAR CONTACT CONDITIONS 3D",
-    }
-
-    def __init__(self, *, cubit=None):
-        """Initialize the input file.
-
-        Args:
-            cubit:
-                A cubit object, that contains an input file which will be loaded
-                into this input file.
-        """
+    def __init__(self):
+        """Initialize the input file."""
 
         # Everything that is not a full MeshPy object is stored here, e.g., parameters
         # and imported nodes/elements/materials/...
@@ -213,131 +93,6 @@ class InputFile:
 
         # Contents of NOX xml file.
         self.nox_xml_contents = ""
-
-        # TODO fix once cubit is converted to YAML
-        # if cubit is not None:
-        #     self._read_dat_lines(cubit.get_dat_lines())
-
-    @classmethod
-    def from_4C_yaml(
-        cls: type[T], input_file_path: _Path, convert_input_to_mesh: bool = False
-    ) -> _Tuple[T, _Mesh]:
-        """Read an existing input file and optionally convert it into a MeshPy
-        mesh.
-
-        Args:
-            input_file_path: A file path to an existing input file that will be read
-                into this object.
-            convert_input_to_mesh: If True, the input file will be converted to a
-                MeshPy mesh.
-
-        Returns:
-            A tuple with the input file and the mesh. If convert_input_to_mesh is
-            False, the mesh will be empty.
-        """
-
-        if _fourcipp_is_available():
-            raise ValueError("Use fourcipp to parse the yaml file.")
-
-        instance = cls()
-
-        with open(input_file_path) as stream:
-            instance.sections = _yaml.safe_load(stream)
-
-        if convert_input_to_mesh:
-            return instance, instance.sections_to_mesh()
-        else:
-            return instance, _Mesh()
-
-    def sections_to_mesh(self):
-        """Convert mesh items, e.g., nodes, elements, element sets, node sets,
-        boundary conditions, materials, ...
-
-        Note: In the current implementation we cannibalize the mesh sections in
-            the self.sections dictionary. This should be reconsidered and be done
-            in a better way when this function is generalized.
-
-        to "true" MeshPy objects.
-        """
-
-        def _get_section_items(section_name):
-            """Return the items in a given section.
-
-            Since we will add the created MeshPy objects to the mesh, we
-            delete them from the plain data storage to avoid having
-            duplicate entries.
-            """
-            if section_name in self.sections:
-                return_list = self.sections[section_name]
-                self.sections[section_name] = []
-            else:
-                return_list = []
-            return return_list
-
-        # Go through all sections that have to be converted to full MeshPy objects
-        mesh = _Mesh()
-
-        # Add nodes
-        for item in _get_section_items("NODE COORDS"):
-            mesh.nodes.append(_Node.from_legacy_string(item))
-
-        # Add elements
-        for item in _get_section_items("STRUCTURE ELEMENTS"):
-            if _fourcipp_is_available():
-                raise ValueError(
-                    "Port this functionality to not use the legacy string."
-                )
-
-            # Get a list containing the element nodes.
-            element_nodes = []
-            for split_item in item.split()[3:]:
-                if split_item.isdigit():
-                    node_id = int(split_item) - 1
-                    element_nodes.append(mesh.nodes[node_id])
-                else:
-                    break
-            else:
-                raise ValueError(
-                    f'The input line:\n"{item}"\ncould not be converted to a element!'
-                )
-
-            mesh.elements.append(_Element.from_legacy_string(element_nodes, item))
-
-        # Add geometry sets
-        geometry_sets_in_sections = {key: None for key in _mpy.geo}
-        for section_name in self.sections.keys():
-            if section_name.endswith("TOPOLOGY"):
-                section_items = _get_section_items(section_name)
-                if len(section_items) > 0:
-                    # Get the geometry key for this set
-                    for key, value in self.geometry_set_names.items():
-                        if value == section_name:
-                            geometry_key = key
-                            break
-                    else:
-                        raise ValueError(f"Could not find the set {section_name}")
-                    geometry_sets_in_section = _get_yaml_geometry_sets(
-                        mesh.nodes, geometry_key, section_items
-                    )
-                    geometry_sets_in_sections[geometry_key] = geometry_sets_in_section
-                    mesh.geometry_sets[geometry_key] = list(
-                        geometry_sets_in_section.values()
-                    )
-
-        # Add boundary conditions
-        for (
-            bc_key,
-            geometry_key,
-        ), section_name in self.boundary_condition_names.items():
-            for item in _get_section_items(section_name):
-                geometry_set_id = item["E"]
-                geometry_set = geometry_sets_in_sections[geometry_key][geometry_set_id]
-                mesh.boundary_conditions.append(
-                    (bc_key, geometry_key),
-                    _boundary_condition_from_dict(geometry_set, bc_key, item),
-                )
-
-        return mesh
 
     def add(self, *args, **kwargs):
         """Add to this object.
@@ -481,12 +236,12 @@ class InputFile:
             """Get the indices for the first "real" MeshPy geometry sets."""
 
             start_indices_geometry_set = {}
-            for geometry_type, section_name in self.geometry_set_names.items():
+            for geometry_type, section_name in _geometry_set_names.items():
                 max_geometry_set_id = 0
                 if section_name in dictionary:
                     section_list = dictionary[section_name]
                     if len(section_list) > 0:
-                        geometry_set_dict = _get_geometry_set_indices_from_section(
+                        geometry_set_dict = get_geometry_set_indices_from_section(
                             section_list, append_node_ids=False
                         )
                         max_geometry_set_id = max(geometry_set_dict.keys())
@@ -657,7 +412,7 @@ class InputFile:
                 section_name = (
                     bc_key
                     if isinstance(bc_key, str)
-                    else self.boundary_condition_names[bc_key, geom_key]
+                    else _boundary_condition_names[bc_key, geom_key]
                 )
                 _dump_mesh_items(self.sections, section_name, bc_list)
 
@@ -668,7 +423,7 @@ class InputFile:
         # Add the geometry sets.
         for geom_key, item in mesh_sets.items():
             if len(item) > 0:
-                _dump_mesh_items(self.sections, self.geometry_set_names[geom_key], item)
+                _dump_mesh_items(self.sections, _geometry_set_names[geom_key], item)
 
         # Add the nodes and elements.
         _dump_mesh_items(self.sections, "NODE COORDS", mesh.nodes)
